refactor: move discover dataset to utils/dataset_samples
